C Code automatic tester&grader
An automatic tester which compiles & checks the results of students' C projects, then grades them accordingly. From an OS course. It gets a config file which includes the path to the projects folder, a path to an input .txt file which is the input for the projects and the path to the expected output .txt file.